changes content of app settings
