+++ src/screens/AppSettings/AppSettings.tsx
@@ -11,11 +11,8 @@ import Twitter from 'src/assets/images/Twitter.svg';
 import Telegram from 'src/assets/images/Telegram.svg';
 import KeeperHeader from 'src/components/KeeperHeader';
 import { LocalizationContext } from 'src/context/Localization/LocContext';
-import LoginMethod from 'src/models/enums/LoginMethod';
 import ScreenWrapper from 'src/components/ScreenWrapper';
 import openLink from 'src/utils/OpenLink';
-import { setSatsEnabled } from 'src/store/reducers/settings';
-import { useAppDispatch, useAppSelector } from 'src/store/hooks';
 import OptionCard from 'src/components/OptionCard';
 import Switch from 'src/components/Switch/Switch';
 import { KEEPER_KNOWLEDGEBASE, KEEPER_WEBSITE_BASE_URL } from 'src/core/config';
@@ -25,18 +22,13 @@ import CurrencyTypeSwitch from 'src/components/Switch/CurrencyTypeSwitch';
 import CircleIconWrapper from 'src/components/CircleIconWrapper';
 
 function AppSettings({ navigation }) {
-  const { colorMode } = useColorMode();
-  const { satsEnabled }: { loginMethod: LoginMethod; satsEnabled: boolean } = useAppSelector(
-    (state) => state.settings
-  );
-
-  const dispatch = useAppDispatch();
+  const { colorMode, toggleColorMode } = useColorMode();
 
   const { translations } = useContext(LocalizationContext);
   const { common, settings } = translations;
 
-  const changeSatsMode = () => {
-    dispatch(setSatsEnabled(!satsEnabled));
+  const changeThemeMode = () => {
+    toggleColorMode();
   };
 
   const actionCardData = [
@@ -59,6 +51,7 @@ function AppSettings({ navigation }) {
     },
   ];
 
+  //TODO: add learn more modal
   return (
     <ScreenWrapper barStyle="dark-content" backgroundcolor={`${colorMode}.primaryBackground`}>
       <KeeperHeader
@@ -93,20 +86,20 @@ function AppSettings({ navigation }) {
           </Box>
         </ScrollView>
         <OptionCard
-          title={settings.SatsMode}
-          description={settings.satsModeSubTitle}
-          callback={() => changeSatsMode()}
+          title={settings.DarkMode}
+          description={settings.DarkModeSubTitle}
+          callback={() => changeThemeMode()}
           Icon={
             <Switch
-              value={satsEnabled}
-              onValueChange={() => changeSatsMode()}
+              onValueChange={() => changeThemeMode()}
+              value={colorMode === 'dark'}
               testID="switch_darkmode"
             />
           }
         />
         <OptionCard
-          title={settings.PrivacyAndDisplay}
-          description={settings.PrivacySettingsSubtitle}
+          title={settings.SecurityAndLogin}
+          description={settings.SecurityAndLoginSubtitle}
           callback={() => navigation.navigate('PrivacyAndDisplay')}
         />
         <OptionCard
@@ -115,18 +108,18 @@ function AppSettings({ navigation }) {
           callback={() => navigation.navigate('NodeSettings')}
         />
         <OptionCard
-          title={settings.VersionHistory}
-          description={settings.VersionHistorySubTitle}
-          callback={() => navigation.navigate('AppVersionHistory')}
-        />
-        <OptionCard
           title={settings.torSettingTitle}
           description={settings.torSettingSubTitle}
           callback={() => navigation.navigate('TorSettings')}
         />
         <OptionCard
-          title={settings.LanguageCountry}
-          description={settings.LanguageCountrySubTitle}
+          title={settings.VersionHistory}
+          description={settings.VersionHistorySubTitle}
+          callback={() => navigation.navigate('AppVersionHistory')}
+        />
+        <OptionCard
+          title={settings.CurrencyDefaults}
+          description={settings.CurrencyDefaultsSubtitle}
           callback={() => navigation.navigate('ChangeLanguage')}
         />
       </ScrollView>
@@ -258,6 +251,7 @@ const styles = StyleSheet.create({
   actionContainer: {
     flexDirection: 'row',
     gap: 5,
+    marginBottom: 20,
   },
   bottomNav: {
     flexDirection: 'row',
